Restructure project and add OpenAPI spec for Servant API.
