+++ Backend/Databases/IMongoDatabaseFactory.cs
@@ -1,0 +1,9 @@
+using MongoDB.Driver;
+
+namespace ZetsubouGacha.Databases
+{
+    public interface IMongoDatabaseFactory
+    {
+        IMongoDatabase Connect(string connectionString, string databaseName);
+    }
+}
