+++ Backend/Servant/Controllers/ServantController.cs
@@ -1,10 +1,9 @@
 ﻿using Microsoft.AspNetCore.Mvc;
-using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using ZetsubouGacha.Databases;
 using ZetsubouGacha.Servants.Models;
-using ZetsubouGacha.Servants.Services;
 
 namespace ZetsubouGacha.Servants.Controllers
 {
