--- Backend/Servant/Models/IMongoDatabaseFactory.cs
@@ -1,10 +1,0 @@
-using System;
-using MongoDB.Driver;
-
-namespace ZetsubouGacha.Servants.Models
-{
-    public interface IMongoDatabaseFactory
-    {
-        IMongoDatabase Connect(string connectionString, string databaseName);  
-    }
-}
