+++ Backend/Services/DbContext.cs
@@ -1,13 +1,13 @@
-using System;
+using ZetsubouGacha.Databases;
 using ZetsubouGacha.Servants.Models;
 
-namespace ZetsubouGacha.Servants.Services
+namespace ZetsubouGacha.Services
 {
     public class DbContext : IDbContext
     {
         public DbContext(IRepositoryFactory repositoryFactory, string connectionString, string dbName)
         {
-            this.Servants = repositoryFactory.Create(new RepositoryOptions(connectionString, dbName, "Servants"));
+            Servants = repositoryFactory.Create(new RepositoryOptions(connectionString, dbName, "Servants"));
         }
 
         public IServantRepository Servants { get; private set; }
