+++ Backend/Startup.cs
@@ -3,9 +3,10 @@ using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
+using ZetsubouGacha.Databases;
 using ZetsubouGacha.Servants.Models;
-using ZetsubouGacha.Servants.Services;
-using ZetsubouGacha.Servants.Settings;
+using ZetsubouGacha.Services;
+using ZetsubouGacha.Settings;
 
 namespace ZetsubouGacha
 {
