+++ Test/Backend.Tests/Backend.Tests.csproj
@@ -16,7 +16,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\Backend\Backend.csproj" />
+    <ProjectReference Include="..\..\Backend\ZetsubouGacha.csproj" />
   </ItemGroup>
 
 </Project>
