+++ zetsubou-gacha.sln
@@ -3,9 +3,14 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.29519.87
 MinimumVisualStudioVersion = 15.0.26124.0
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Backend", "Backend\Backend.csproj", "{259224F5-5843-4A82-B5D6-D5C885E1CBBD}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ZetsubouGacha", "Backend\ZetsubouGacha.csproj", "{259224F5-5843-4A82-B5D6-D5C885E1CBBD}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Backend.Tests", "Test\Backend.Tests\Backend.Tests.csproj", "{EF3FFA74-6695-48E1-85A9-E9607B0F309A}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Backend.Tests", "Test\Backend.Tests\Backend.Tests.csproj", "{EF3FFA74-6695-48E1-85A9-E9607B0F309A}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{123E27A2-D2D9-4E69-96D7-960785E2F5BF}"
+	ProjectSection(SolutionItems) = preProject
+		Servant.v1.yaml = Servant.v1.yaml
+	EndProjectSection
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
